requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.4
fastapi-cli==0.0.5
fonttools==4.55.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
ipywidgets==8.1.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
numpy==2.1.2
oauthlib==3.2.2
openai==1.55.1
orjson==3.10.11
outcome==1.3.0.post0
pandas==2.2.3
pillow==11.0.0
playwright==1.49.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyee==12.0.0
pyparsing==3.2.0
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.17
pytrends==4.9.2
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
selenium==4.25.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.2
tiktoken==0.8.0
tqdm==4.66.5
trio==0.27.0
trio-websocket==0.11.1
typer==0.12.5
tzdata==2024.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
widgetsnbextension==4.0.13
wsproto==1.2.0
youtube-transcript-api==0.6.2
colorama==0.4.6
